[Fix] zustand stor 파일 분리로 인한 경로 수정
